requirements.txt
lxml==4.3.3
python-docx==0.8.10
tqdm==4.32.1
